Add start_delay and toast once functionality
Start_delay will delay the initial start of node-dev.  This is useful
if used with a tool like 'tsc -w' and node-dev starts while tsc is
still compiling files.

Toast once will prevent getting spammed if the same file is changed.
This may only be an issue on Windows, but for some reason if a file is
compiled with tsc, node-dev receives multiple events for it.